+++ packages/x-components/src/composables/index.ts
@@ -3,3 +3,5 @@ export * from './use-$x';
 export * from './use-register-x-module';
 export * from './use-on-display';
 export * from './use-store';
+export * from './use-state';
+export * from './use-getter';