+++ packages/x-components/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0-alpha.18](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.17...@empathyco/x-components@4.1.0-alpha.18) (2024-02-16)
+
+
+### Code Refactoring
+
+* **components:** migrate animate-width component to vue 2.7 (#1411) ([47c26fb](https://github.com/empathyco/x/commit/47c26fba64025ae6fc391979e8d82fd366afe5b9))
+
+
+
+## [4.1.0-alpha.17](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.16...@empathyco/x-components@4.1.0-alpha.17) (2024-02-14)
+
+
+### Features
+
+* **composables:** create use getter composable (#1407) ([44a1488](https://github.com/empathyco/x/commit/44a148848ab027a624a3a1980ed4a62cb0956dea))
+
+
+
+## [4.1.0-alpha.16](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.15...@empathyco/x-components@4.1.0-alpha.16) (2024-02-14)
+
+
+### Testing
+
+* uncomment tests that were expected to fail (#1405) ([f2310a4](https://github.com/empathyco/x/commit/f2310a43133a1b6e1531d59d8e392912eff96894))
+
+
+
+## [4.1.0-alpha.15](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.14...@empathyco/x-components@4.1.0-alpha.15) (2024-02-13)
+
+
+### Features
+
+* **composables:** create useState composable (#1402) ([435af42](https://github.com/empathyco/x/commit/435af4222a2ed08d44d728a973406c0e1122b862))
+
+
+
 ## [4.1.0-alpha.14](https://github.com/empathyco/x/compare/@empathyco/x-components@4.1.0-alpha.13...@empathyco/x-components@4.1.0-alpha.14) (2024-02-05)